+++ editor/src/components/canvas/design-panel-root.tsx
@@ -36,9 +36,9 @@ import { InsertMenuPane } from '../navigator/insert-menu-pane'
 import { CanvasToolbar } from '../editor/canvas-toolbar'
 import { useDispatch } from '../editor/store/dispatch-context'
 import { LeftPaneComponent } from '../navigator/left-pane'
-import type { StoredPanel } from './grid-panels-container'
+import type { StoredPanel } from './grid-panels-state'
 import { GridPanelsContainer } from './grid-panels-container'
-import type { Menu, Pane } from './grid-panels-state-2'
+import type { Menu, Pane } from './grid-panels-state'
 import type { ResizableProps } from '../../uuiui-deps'
 import type { Direction } from 're-resizable/lib/resizer'
 import { isFeatureEnabled } from '../../utils/feature-switches'
