+++ editor/src/components/titlebar/title-bar.tsx
@@ -29,7 +29,7 @@ import { Substores, useEditorState } from '../editor/store/store-hook'
 import { RoundButton } from './buttons'
 import { TestMenu } from './test-menu'
 import { useGridPanelDraggable } from '../canvas/grid-panels-dnd'
-import type { StoredPanel } from '../canvas/grid-panels-container'
+import type { StoredPanel } from '../canvas/grid-panels-state'
 
 interface ProjectTitleProps {}
